+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added
 
+- Cerebras language model provider
 - Stream audio setting to Picovoice/Deepgram STT providers
 - Open files to edit mode prompt